将根据商品id查找商品评论修改为根据店铺id查找评论。

+++ src/main/java/com/pro516/thrifttogetherbackstage/entity/vo/SimpleReviewVO.java
@@ -36,7 +36,7 @@ public class SimpleReviewVO {
     private List<String> reviewPicUrlList;
 
     @ApiModelProperty(value = "评价分数")
-    private Double reviewScore;
+    private Integer reviewScore;
 
     @JsonFormat(pattern = "yyyy-MM-dd")
     @ApiModelProperty(value = "评价时间")
